Add the ABI lib to handle encoding and decoding
This lib makes it easy to handle multiple types of arguments and it
also deals with all the Keccak256 cryptography. The need to handle
multiple types of arguments comes from the Ethereum Virtual Machine,
that expects those values in a particular way. Also, the functions'
names should also be written in a particular way using the Keccak256
algorithm.